winter_prover

Trait ConstraintCommitment

pub trait ConstraintCommitment<E: FieldElement> {
    type HashFn: ElementHasher<BaseField = E::BaseField>;
    type VC: VectorCommitment<Self::HashFn>;

    // Required methods
    fn commitment(&self) -> <Self::HashFn as Hasher>::Digest;
    fn query(self, positions: &[usize]) -> Queries;
}

Constraint evaluation commitment.

The commitment consists of two components:

Evaluations of composition polynomial columns over the LDE domain.

Vector commitment where each vector element corresponds to the digest of a row in the composition polynomial evaluation matrix.

Required Associated Types

type HashFn: ElementHasher<BaseField = E::BaseField>

The hash function used for hashing the rows of trace segment LDEs.

type VC: VectorCommitment<Self::HashFn>

The vector commitment scheme used for commiting to the trace.

Required Methods

fn commitment(&self) -> <Self::HashFn as Hasher>::Digest

Returns the commitment.

fn query(self, positions: &[usize]) -> Queries

Returns constraint evaluations at the specified positions along with a batch opening proof against the vector commitment.