winter_prover

Trait Trace

pub trait Trace: Sized {
    type BaseField: StarkField;

    // Required methods
    fn info(&self) -> &TraceInfo;
    fn main_segment(&self) -> &ColMatrix<Self::BaseField>;
    fn read_main_frame(
        &self,
        row_idx: usize,
        frame: &mut EvaluationFrame<Self::BaseField>,
    );

    // Provided methods
    fn length(&self) -> usize { ... }
    fn main_trace_width(&self) -> usize { ... }
    fn aux_trace_width(&self) -> usize { ... }
    fn validate<A, E>(
        &self,
        air: &A,
        aux_trace_with_metadata: Option<&AuxTraceWithMetadata<E, <A as Air>::GkrProof>>,
    )
       where A: Air<BaseField = Self::BaseField>,
             E: FieldElement<BaseField = Self::BaseField> { ... }
}

Defines an execution trace of a computation.

Execution trace can be reduced to a two-dimensional matrix in which each row represents the state of a computation at a single point in time and each column corresponds to an algebraic column tracked over all steps of the computation.

Building a trace is required for STARK proof generation. An execution trace of a specific instance of a computation must be supplied to Prover::prove() method to generate a STARK proof.

This crate exposes one concrete implementation of the Trace trait: TraceTable. This implementation supports concurrent trace generation and should be sufficient in most situations. However, if functionality provided by TraceTable is not sufficient, uses can provide custom implementations of the Trace trait which better suit their needs.

Required Associated Types

type BaseField: StarkField

Base field for this execution trace.

All cells of this execution trace contain values which are elements in this field.

Required Methods

fn info(&self) -> &TraceInfo

Returns trace info for this trace.

fn main_segment(&self) -> &ColMatrix<Self::BaseField>

Returns a reference to a [Matrix] describing the main segment of this trace.

fn read_main_frame( &self, row_idx: usize, frame: &mut EvaluationFrame<Self::BaseField>, )

Reads an evaluation frame from the main trace segment at the specified row.

Provided Methods

fn length(&self) -> usize

Returns the number of rows in this trace.

fn main_trace_width(&self) -> usize

Returns the number of columns in the main segment of this trace.

fn aux_trace_width(&self) -> usize

Returns the number of columns in the auxiliary trace segment.

fn validate<A, E>( &self, air: &A, aux_trace_with_metadata: Option<&AuxTraceWithMetadata<E, <A as Air>::GkrProof>>, )
where A: Air<BaseField = Self::BaseField>, E: FieldElement<BaseField = Self::BaseField>,

Checks if this trace is valid against the specified AIR, and panics if not.

NOTE: this is a very expensive operation and is intended for use only in debug mode.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.